Use properties 'minor', 'micro', 'reqMinor' instead of 'version'.
On Linux, version number is appended to 'so' files, when using 'version'
property. This causes problems as Qbs doesn't create symbolic links in
install directory (it does so in build directory). There's 'cpp.soVersion'
property, but it doesn't seem to work as described. For now 'version'
property won't be used. Specifying 'minor' and 'reqMinor' explicitly seems
to be more clear anyways.