Improve units.
Property 'strokeWidth' width is now binded to 'strokeWidthRatio'.

Indirection for 'units' is introduced in Element, so that custom Units
can be provided for each element without the need of modifying whole theme.
It also helps to deal with Qt Creator's Designer, which has troubles with
non-static content, such as when theme is updated from
'Component.onCompleted' handler.

Affected components are updated accordingly.