Use persistent states for transitions
The State Machine Framework's behavior is to always enter parent state
first, even if transition defines child as a target state. This may cause
problems when trying to trigger the transition after entering the
persistent state. Since ephemeric child may not be entered yet the state
machine may fail to trigger the transition, when the transition is defined
for the ephemeric child rather than the persistent parent state.