Prepend metadata path with QCoreApplication::applicationDirPath()
When the application runs from a bundle such as AppImage it will fail to
load the file if the actual path to the resource inside the bundle is not
provided. Without QCoreApplication::applicationDirPath() a relative path
is returned. Functions like QFile::open() take relative path as relative
to the location of the bundle and not its mounted contents, where the
executable and metadata really is.